Added Preview component
Added a new Preview component that renders the Emoji Card without the editing interface.
Isolated the logic that iterates and renders the Emojis from the stored array on its own
component, avoiding code repetition. Implemented conditional rendering based on props in
components with editing interfaces, hiding them depending on which view is being used.
