+++ src/App.vue
@@ -1,6 +1,7 @@
 <template>
   <div id="nav">
     <router-link to="/">Home</router-link> |
+    <router-link to="/preview">Preview</router-link> |
     <router-link to="/about">About</router-link>
   </div>
   <router-view />
