+++ src/views/Home.vue
@@ -2,36 +2,13 @@
   <div class="home">
     <Message
       :msg="state.message1"
+      :editMode="true"
       @update:newMessage="state.message1 = $event"
     />
-    <div class="emoji-box-wrapper">
-      <div
-        class="emoji-wrapper"
-        v-for="(codepoint, index) in state.activeEmojis"
-        :key="index"
-      >
-        <Emoji
-          class="emoji-box"
-          :position="index"
-          :codepoint="codepoint"
-          :label="state.allEmojis.get(codepoint).label"
-        />
-        <div class="button-wrapper">
-          <button class="button-left" @click="moveLeft(index)">&#60;</button>
-          <button
-            @click="shiftEmoji(index, state.allEmojis.get(codepoint).category)"
-          >
-            Shift
-          </button>
-          <button @click="removeEmoji(index)">Remove</button>
-          <button class="button-right" @click="moveRight(index)">
-            &#62;
-          </button>
-        </div>
-      </div>
-    </div>
+    <EmojiWrapper :editMode="true" />
     <Message
       :msg="state.message2"
+      :editMode="true"
       @update:newMessage="state.message2 = $event"
     />
     <button class="button-left" @click="setDefault()">Start over</button>
@@ -43,27 +20,21 @@
 </template>
 
 <script lang="ts">
-import { defineComponent, onMounted } from "vue";
+import { defineComponent } from "vue";
 
 import state from "@/store/state";
 import { ValidCategories } from "@/assets/types";
-import {
-  loadEmojis,
-  setDefault,
-  setEmoji,
-  shiftEmoji,
-  removeEmoji
-} from "@/store/methods";
+import { shiftEmoji, setDefault } from "@/store/methods";
 
 import Message from "@/components/Message.vue";
-import Emoji from "@/components/Emoji.vue";
+import EmojiWrapper from "@/components/EmojiWrapper.vue";
 import AddEmoji from "@/components/AddEmoji.vue";
 
 export default defineComponent({
   name: "Home",
   components: {
     Message,
-    Emoji,
+    EmojiWrapper,
     AddEmoji
   },
   data() {
@@ -72,57 +43,12 @@ export default defineComponent({
     };
   },
   setup() {
-    onMounted(() => {
-      if (state.allEmojis.size === 0) {
-        loadEmojis();
-      }
-      if (state.activeEmojis.length === 0) {
-        setDefault();
-      }
-      console.log(state.activeEmojis);
-      console.log(`Loaded emojis: ${state.allEmojis.size.toString()} \n`);
-    });
-
     function newEmoji(category: ValidCategories) {
       console.log(category);
       shiftEmoji(state.activeEmojis.length, category);
     }
 
-    function moveLeft(position: number) {
-      console.log("Moving left from position", position);
-      if (position > 0) {
-        const currentEmoji: string | undefined = state.activeEmojis[position];
-        const leftEmoji: string | undefined = state.activeEmojis[position - 1];
-
-        if (currentEmoji != undefined && leftEmoji != undefined) {
-          setEmoji(position - 1, currentEmoji);
-          setEmoji(position, leftEmoji);
-        }
-      }
-    }
-
-    function moveRight(position: number) {
-      console.log("Moving right from position", position);
-      if (position < state.activeEmojis.length) {
-        const currentEmoji: string | undefined = state.activeEmojis[position];
-        const rightEmoji: string | undefined = state.activeEmojis[position + 1];
-
-        if (currentEmoji != undefined && rightEmoji != undefined) {
-          setEmoji(position + 1, currentEmoji);
-          setEmoji(position, rightEmoji);
-        }
-      }
-    }
-
-    return {
-      onMounted,
-      shiftEmoji,
-      newEmoji,
-      removeEmoji,
-      setDefault,
-      moveLeft,
-      moveRight
-    };
+    return { newEmoji, setDefault };
   }
 });
 </script>
@@ -131,31 +57,6 @@ export default defineComponent({
 .home {
   border: 1px solid black;
   font-family: "Verdana", "Geneva", sans-serif;
-}
-.emoji-box-wrapper {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-}
-.emoji-wrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  margin-bottom: 1rem;
-  border: 1px solid black;
-}
-.emoji-box {
-  width: 12rem;
-  height: 12rem;
-  margin: 1rem;
-  font-size: 9.5rem;
-  border: 1px dashed black;
-}
-.button-wrapper {
-  flex-direction: row;
-  justify-content: space-between;
-  border: 1px solid black;
 }
 button {
   margin: 0;
@@ -181,7 +82,7 @@ button:focus {
   border-top-right-radius: 5em;
   border-bottom-right-radius: 5em;
 }
-
+/* 
 @media only screen and (max-width: 600px) {
   .upper-wrapper {
     flex-direction: column;
@@ -192,5 +93,5 @@ button:focus {
     flex-direction: row;
     justify-content: space-around;
   }
-}
+} */
 </style>
